requirements.txt
streamlit
openai
pandas
pillow
speechrecognition
transformers
torch
